requirements.txt
jax==0.3.13
jaxlib==0.3.10+cuda11.cudnn82
flax==0.5.2
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-text==2.9.0
PyWavelets==1.3.0
absl-py==1.1.0